feat(shared): add Flyway Migrations and expose in ZIO Layer to be used in consuming service
Previously, migrations were managed by the infrastructure layer and needed to be composed at docker image build stage.
This change allows a consumer of the pollux and castor library to execute migrations in the application layer, by composing a
db config layer with the provided miggrations layer and calling the `migrate` method on the output.

Implements ATL-2174