feat(shared): set OAS servers to `k8s-dev.atalaprism.io` and add API Key auth method - also update local docker-compose implementation (#126)
Previsouly, no authentication mechanism was specified on the OAS files served by prism-agent and mediator. Now, the OAS contains
instructions on how to authenticate to the APIs as well as a change to the servers to make sure requests go to a real running instance.
In the same PR, the local docker-compose implementation has been updated to be in line with the running environmnet. Please note
the changes to the mercury-mediator routes - rather than the internal service name being exposed - it has been changed to `mediator`. This
makes more sense to external consumers when they try to understand what the service is and does. Also, the names of images
and the docker repository has been updated to reflect CI changes and the fact we are now storing docker images in Github Packages.
Finally, a number of local helper scripts have been added to simplify local development [clean and build and run-only]

Implements ATL-2215