Update docker-compose.yml
Restricting postgres and pgadmin to local interface by default seems to be a better option since there is no reason to expose this ports to the public internet. Anyone following the local deployment guide on a publicly accesible web server it's a high risk of having the container hijacked by malware bots when the ports are exposed on every interface.

Signed-off-by: Roberto Carvajal <roberto.carvajal@gmail.com>